feat(table): fix query to fetch from correct table in graphql
